Move some stuff to util.js
move toString() now works
